change: Move Compute_sum_Mz into util mode
This method was part of CCCS associated methods. But it doesn't make
sense as LCCCS also requires it aside from other standalone functions.

Hence, it is moved to a util module.
